requirements.txt
Django==2.0.5; python_version >= '3.0' 
Django==1.7.5; python_version < '3.0' 
Pillow
MySQL-python; python_version < '3.0' 
mysqlclient; python_version >= '3.0'  
django-crispy-forms==1.7.2
django-registration-redux==2.4